refactor: change folder structure to fit convention
